requirements.txt
# requirements.txt

# Common & Frontend
streamlit==1.35.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0

# Backend
flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# NLP and Text Processing
spacy==3.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
PyMuPDF==1.23.8
python-docx==1.1.0
phonenumbers==8.13.27

# High-Performance Semantic Search & Embeddings
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.35.2

# AI Model for Summarization
google-generativeai==0.3.2